pyproject.toml
[project]
name = "chemview"
version = "0.1.0"
description = "CLI for visualizing SMILES strings as 2D molecule images"
readme = "README.md"
authors = [
    { name = "Raymond Gasper", email = "raymondgasper@fastmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "typer>=0.9",
    "rich>=13",
    "rdkit>=2024.3",
    "Pillow>=10",
    "loguru>=0.7",
]

[project.scripts]
chemview = "chemview.cli:app"

[dependency-groups]
dev = [
    "pytest>=8",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
